renamed folder and removed references to it
